+++ build_linux.sh
@@ -38,6 +38,7 @@ EXTRA=()
     --hidden-import urllib.error \
     --hidden-import configparser \
     --hidden-import concurrent.futures \
+    --hidden-import certifi \
     --exclude-module pytest \
     "${EXTRA[@]}" \
     main.py